cmd/k8s-operator: add CRD to static manifest
The static manifest that folks use to kubectl apply operator resources
is generated from Helm templates.
Update the generation logic to ensure that the Connector CRD too
gets added to the static manifest. This is so that folks who use
the static manifest don't need to run two commands to install
now that we have a CRD. Duplicating the CRD is suboptimal
and a long term solution should be to generate and publish a static manifest
- either to Github releases or packages server.

Updates tailscale/tailscale#10641

Signed-off-by: Irbe Krumina <irbe@tailscale.com>