cmd/k8s-operator: annotate proxy StatefulSets with operator version
Add a tailscale.com/operator-last-version annotation to StatefulSets for ingress/egress proxies and Connectors.
Set it to the operator version of the current operator version each time the StatefulSet is re-synced.
This will help us to determine the potential proxy state when making changes in proxy configuration in the future.

Updates tailscale/tailscale#10407

Signed-off-by: Irbe Krumina <irbe@tailscale.com>